% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_data_task.R
\docType{data}
\name{data_task}
\alias{data_task}
\title{Dataset, Multi assay data in mlr's task format}
\format{mlr's task
\describe{
  \item{data_task}{task}
  ...
}}
\source{
JDRF multi center efforts
}
\usage{
data_task
}
\description{
mlr's task (NO functionals implemented YET), therefore, use sep to at analyte name to exatract assay annotation
}
\keyword{datasets}
