% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_pre_process_helpers.R
\name{raw_pre_process_helpers}
\alias{raw_pre_process_helpers}
\alias{store_All_list_into_MAE}
\alias{Setup.plue.MAE.to.vertical.function}
\alias{Insert.Metadata.Assay.type.FUNCTION}
\alias{MAE.to.Join_SumEset.FUNCTION}
\alias{Screen.DF.Standarize.by.assay.FUNCTION}
\title{raw data pre-processing functions}
\usage{
store_All_list_into_MAE(All.list.raw.only, pData.in)

Setup.plue.MAE.to.vertical.function(MAE, param.Y.name, param.assays.vector,
  param.subjects.study, param.cohort, param.covariates,
  param.rv144.mRNA.collapse.multiple.probes, parame.gene.or.module,
  param.Selected.Meta.sets, Assay.Analyte.sep)

Insert.Metadata.Assay.type.FUNCTION(SumEset.i, Assay, Assay.type)

MAE.to.Join_SumEset.FUNCTION(MAE, Assay.Analyte.sep)

Screen.DF.Standarize.by.assay.FUNCTION(DF, Assay.Analyte.sep)
}
\arguments{
\item{all_kind}{all kind...}
}
\value{
all kind...
}
\description{
raw data pre-processing functions
}
\details{
TBA
}
